# pkg/api/rest/docs/swagger.yaml
basePath: /
definitions:
  github_com_cloud-barista_cm-centipede_dmdl_common-model.BeetleDBRef:
    properties:
      nsId:
        type: string
      password:
        description: AES-encrypted on store
        type: string
      rdbmsId:
        type: string
      tlsCAPem:
        description: |-
          TLSCAPEM is the PEM root that verify-ca and verify-full check against.
          Empty means the system trust store, which carries only publicly trusted
          CAs — and managed RDBMS services commonly sign with their own, so those two
          modes usually need the CSP's root supplied here.
        type: string
      tlsMode:
        description: |-
          TLSMode selects transport security; the empty value is "disable". It means
          the same thing here as on DBConnConfig — see the table in transx-ex's
          dbmsx/driver/base/tls.go.
        enum:
        - disable
        - prefer
        - require
        - verify-ca
        - verify-full
        type: string
      username:
        description: defaults to the looked-up AdminUsername
        type: string
    required:
    - nsId
    - password
    - rdbmsId
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.BeetleOSRef:
    properties:
      nsId:
        type: string
      osId:
        type: string
    required:
    - nsId
    - osId
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.BeetleSSHRef:
    properties:
      infraId:
        type: string
      nodeId:
        type: string
      nsId:
        type: string
    required:
    - infraId
    - nodeId
    - nsId
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef:
    properties:
      beetleDb:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.BeetleDBRef'
      beetleObjectStorage:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.BeetleOSRef'
      beetleSsh:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.BeetleSSHRef'
      db:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.DBConnConfig'
      honeybee:
        allOf:
        - $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.HoneybeeRef'
        description: Reference types
      minio:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.MinioConnConfig'
      source:
        allOf:
        - $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionSource'
        enum:
        - honeybee
        - beetleSsh
        - beetleObjectStorage
        - beetleDb
        - ssh
        - minio
        - db
      ssh:
        allOf:
        - $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.SSHConnConfig'
        description: Inline types
    required:
    - source
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionSource:
    enum:
    - honeybee
    - beetleSsh
    - beetleObjectStorage
    - beetleDb
    - ssh
    - minio
    - db
    type: string
    x-enum-varnames:
    - ConnectionSourceHoneybee
    - ConnectionSourceBeetleSSH
    - ConnectionSourceBeetleObjectStorage
    - ConnectionSourceBeetleDB
    - ConnectionSourceSSH
    - ConnectionSourceMinio
    - ConnectionSourceDB
  github_com_cloud-barista_cm-centipede_dmdl_common-model.DBConnConfig:
    properties:
      accessType:
        enum:
        - direct
        - sshTunnel
        type: string
      authSource:
        description: |-
          AuthSource selects the authentication database. MongoDB only; ignored by
          other drivers. When empty and credentials are given, MongoDB defaults to
          "admin" (where users are usually created) rather than the target database.
        type: string
      connectTimeout:
        type: integer
      dbmsType:
        enum:
        - mysql
        - mariadb
        - postgresql
        - mongodb
        type: string
      host:
        description: |-
          DB endpoint (the direct connection target, or the target reached through
          the SSH tunnel)
        type: string
      password:
        type: string
      port:
        type: integer
      providerName:
        description: |-
          ProviderName records where the database is hosted. transx-ex carries it
          through to DBMSLocation.ProviderName but never reads it: no routing,
          validation or connection decision depends on it. It is required all the
          same, so that "where does this data live" has an answer for every
          connection that holds data. A database the operator runs themselves is
          ProviderOnPrem, spelled out rather than left blank.
        enum:
        - aws
        - alibaba
        - tencent
        - ncp
        - nhn
        - ibm
        - gcp
        - kt
        - azure
        - openstack
        - onprem
        type: string
      sshTunnel:
        allOf:
        - $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.SSHConnConfig'
        description: Jump host used when accessType=sshTunnel (inline SSH).
      tlsCAPem:
        description: |-
          TLSCAPem is the PEM root that verify-ca and verify-full check against.
          Empty means the system trust store, which carries only publicly trusted
          CAs, so a server using a private CA needs its root supplied here.

          transx-ex also accepts a path to a CA file, but that path would name a
          file on the centipede server rather than on the caller's machine, so it
          is deliberately not part of this API.
        type: string
      tlsMode:
        description: |-
          TLSMode selects transport security; the empty value is "disable". The
          modes mean the same thing on every engine — see transx-ex's
          dbmsx/driver/base/tls.go for the table that defines them.
        enum:
        - disable
        - prefer
        - require
        - verify-ca
        - verify-full
        type: string
      username:
        type: string
    required:
    - accessType
    - dbmsType
    - providerName
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.DBMSFilter:
    properties:
      databases:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.DBMSFilterEntry'
        type: array
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.DBMSFilterEntry:
    properties:
      rules:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.DBMSFilterRule'
        type: array
      targetMapping:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.TargetMapping'
    required:
    - targetMapping
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.DBMSFilterRule:
    properties:
      columns:
        description: type=table_column (empty => whole table)
        items:
          type: string
        type: array
      kind:
        description: type=object_exclude (view|index|function|...)
        type: string
      match:
        description: type=row_exclude
        type: string
      name:
        description: type=object_exclude
        type: string
      table:
        type: string
      type:
        type: string
    type: object
...
  github_com_cloud-barista_cm-centipede_dmdl_common-model.FileSystemFilter:
    properties:
      rules:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.PathFilterRule'
        type: array
      targetMapping:
        allOf:
        - $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.TargetMapping'
        description: selects the single folder to migrate
    required:
    - targetMapping
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.HoneybeeRef:
    properties:
      connectionId:
        type: string
    required:
    - connectionId
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.MinioConnConfig:
    properties:
      accessKeyId:
        description: |-
          AccessKeyId doubles as the storage account name on azure, where it is also
          the endpoint source.
        type: string
      bucketLookup:
        description: '"" | "dns" | "path"'
        type: string
      endpoint:
        description: |-
          Endpoint is required for the openstack and onprem providers, where the
          user supplies the host. For the others it is an optional override of the
          derived host — except azure, whose host is derived from AccessKeyId.
        type: string
      providerName:
        description: |-
          ProviderName decides the endpoint, region, TLS and bucket addressing style
          (see ResolveS3Endpoint in core/migration). transxex.S3MinioConfig has no
          provider field: it receives the resolved values.
        enum:
        - aws
        - alibaba
        - tencent
        - ncp
        - nhn
        - ibm
        - gcp
        - kt
        - azure
        - openstack
        - onprem
        type: string
      region:
        description: |-
          Region is the region this store lives in. Required for the providers whose
          hostname contains it (aws, alibaba, tencent, ncp, nhn, ibm) and for gcp,
          which has a region-less host but still signs with a region.

          It is NOT unconditionally the signing region: NCP serves
          "kr.object.ncloudstorage.com" and signs "kr-standard", so ResolveS3Endpoint
          signs with this value only where the two are spelled the same and otherwise
          leaves the signing region to be discovered.
        type: string
      secretAccessKey:
        type: string
      useSSL:
        description: |-
          UseSSL is honored only for user-supplied endpoints (openstack, onprem).
          The other providers use the fixed value from the provider table.
        type: boolean
    required:
    - accessKeyId
    - providerName
    - secretAccessKey
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.ObjectStorageFilter:
    properties:
      rules:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.PathFilterRule'
        type: array
      targetMapping:
        allOf:
        - $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.TargetMapping'
        description: selects the single bucket to migrate
    required:
    - targetMapping
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.PathFilterRule:
    properties:
      action:
        description: include | exclude
        type: string
      op:
        description: type=size (> >= < <= ==)
        type: string
      pattern:
        description: type=glob
        type: string
      type:
        description: glob | size
        type: string
      value:
        description: type=size (bytes)
        type: integer
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.PgSchemaMapping:
    properties:
      dstName:
        type: string
      srcName:
        type: string
    required:
    - srcName
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.SSHConnConfig:
    properties:
      host:
        type: string
      port:
        description: defaults to 22
        type: integer
      privateKey:
        type: string
      username:
        type: string
    required:
    - host
    - privateKey
    - username
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_common-model.TargetMapping:
    properties:
      dstName:
        type: string
      pgSchemas:
        description: |-
          PgSchemas selects and renames PostgreSQL schemas within this database.
          Empty migrates every schema under its own name; a non-empty list migrates
          only the schemas it names, which makes it a selection as much as a rename.
          PostgreSQL only: the plan layer rejects it for other engines and for the
          file system and object storage filters.
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.PgSchemaMapping'
        type: array
      srcName:
        type: string
    required:
    - srcName
    type: object
...
  github_com_cloud-barista_cm-centipede_dmdl_source-model.DBMSInfo:
    properties:
      characterSet:
        description: |-
          CharacterSet is the database's default character set, which on PostgreSQL
          is its encoding — the two name the same thing there. MongoDB stores UTF-8
          only and leaves it empty.
        type: string
      collation:
        description: |-
          Collation is the database's default collation: DEFAULT_COLLATION_NAME on
          MySQL and MariaDB, datcollate on PostgreSQL. MongoDB has no database-level
          collation and leaves it empty.
        type: string
      ctype:
        description: |-
          Ctype is the locale governing character classification — PostgreSQL's
          datctype, which the engine keeps separate from the collation order. The
          other engines have no equivalent and leave it empty.
        type: string
      dataSize:
        type: integer
      database:
        type: string
      events:
        description: MySQL / MariaDB
        items:
          type: string
        type: array
      extensions:
        description: PostgreSQL
        items:
          type: string
        type: array
      foreignKeys:
        description: Schema object inventory (name list) — populated only on metric
          opt-in.
        items:
          type: string
        type: array
      functions:
        items:
          type: string
        type: array
      indexSize:
        type: integer
      materializedViews:
        description: PostgreSQL
        items:
          type: string
        type: array
      pgSchema:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_source-model.DBPgSchemaInfo'
        type: array
      procedures:
        items:
          type: string
        type: array
      rules:
        description: PostgreSQL
        items:
          type: string
        type: array
      sequences:
        description: PostgreSQL
        items:
          type: string
        type: array
      serverVersion:
        type: string
      tableCount:
        type: integer
      tables:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_source-model.DBTableInfo'
        type: array
      totalSize:
        description: |-
          TotalSize..TableCount are the sum over the schemas the inspection covered,
          not the size of the database as PostgreSQL reports it. PgSchema breaks them
          down per schema and always sums back to them; a schema holding no tables is
          listed at zero rather than omitted. PostgreSQL only.
        type: integer
      triggers:
        items:
          type: string
        type: array
      types:
        description: PostgreSQL
        items:
          type: string
        type: array
      validators:
        description: MongoDB
        items:
          type: string
        type: array
      views:
        items:
          type: string
        type: array
    type: object
...
  github_com_cloud-barista_cm-centipede_dmdl_source-model.SourceDBModel:
    properties:
      connection:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef'
      databases:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_source-model.DBMSInfo'
        type: array
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_source-model.SourceDataMigrationModel:
    properties:
      sourceDataMigrationModel:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_source-model.SourceDataMigrationProperty'
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_source-model.SourceDataMigrationProperty:
    properties:
      databases:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_source-model.SourceDBModel'
        type: array
      fileSystems:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_source-model.SourceFileSystemModel'
        type: array
      objectStorages:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_source-model.SourceObjectStorageModel'
        type: array
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_source-model.SourceFileSystemModel:
    properties:
      connection:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef'
      extensionCounts:
        additionalProperties:
          type: integer
        description: only when FilesystemMetric.ExtensionCount
        type: object
      fileCount:
        description: only when FilesystemMetric.FileCount
        type: integer
      folders:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_source-model.FSEntry'
        type: array
      isMount:
        type: boolean
      path:
        type: string
      totalSize:
        description: only when FilesystemMetric.TotalSize
        type: integer
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_source-model.SourceObjectStorageModel:
    properties:
      connection:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef'
      extensionCounts:
        additionalProperties:
          type: integer
        description: only when ObjectStorageMetric.ExtensionCount
        type: object
      folders:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_source-model.ObjectEntry'
        type: array
      objectCount:
        description: only when ObjectStorageMetric.ObjectCount
        type: integer
      path:
        type: string
      totalSize:
        description: only when ObjectStorageMetric.TotalSize
        type: integer
    type: object
...
  github_com_cloud-barista_cm-centipede_dmdl_target-model.MigrationDBModel:
    properties:
      databases:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_target-model.DBMigrationInfo'
        minItems: 1
        type: array
      dbType:
        allOf:
        - $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.DBMSType'
        enum:
        - mysql
        - mariadb
        - postgresql
        - mongodb
      dstConnection:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef'
      errors:
        items:
          type: string
        type: array
      planEntryId:
        description: |-
          PlanEntryID names this entry within the plan
          — see MigrationFileSystemModel.PlanEntryID.
        type: string
      srcConnection:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef'
    required:
    - databases
    - dbType
    - dstConnection
    - srcConnection
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_target-model.MigrationFileSystemModel:
    properties:
      dstConnection:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef'
      dstType:
        description: |-
          DstType is the destination storage type: empty/"filesystem" = SSH→SSH,
          "objectstorage" = filesystem→objectstorage cross-storage.
        enum:
        - filesystem
        - objectstorage
        type: string
      errors:
        items:
          type: string
        type: array
      folders:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_target-model.FSMigrationInfo'
        type: array
      planEntryId:
        description: |-
          PlanEntryID names this entry within the plan. It is what a migration log,
          a validation detail and a per-entry retry refer to, so that none of them
          has to identify an entry by its position in the array — a position is not
          a name, and nothing outside the plan can hold on to one.

          "entry", not "item": an item is the folder, bucket or database inside one
          of these (see countPlanItems, MigrationLog.ItemPath, Migration.TotalItems),
          and a name that meant both layers at once would be worse than the index it
          replaces. Not "id" either — an entry is anonymous in its own document,
          sitting in fileSystems[] with no type name beside it, so "id" alone would
          not say what kind of id it is. The log field that refers to it is spelled
          the same, which is the point: one token to search for in both documents.

          Assigned by POST /plans/target. A plan that arrives without one — edited,
          hand-written, issued elsewhere — is given ids by CreateMigration before it
          is stored, so a stored plan always has them.
        type: string
      srcConnection:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef'
      strategy:
        description: |-
          Strategy is the transx transfer strategy (auto/direct/relay). It only affects
          SSH→SSH transfers; cross-storage always relays. Set by the plan layer.
        enum:
        - auto
        - direct
        - relay
        type: string
    required:
    - dstConnection
    - srcConnection
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_target-model.MigrationObjectStorageModel:
    properties:
      buckets:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_target-model.ObjectMigrationInfo'
        type: array
      dstConnection:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef'
      dstType:
        description: |-
          DstType is the destination storage type: empty/"objectstorage" = S3→S3,
          "filesystem" = objectstorage→filesystem cross-storage.
        enum:
        - filesystem
        - objectstorage
        type: string
      errors:
        items:
          type: string
        type: array
      planEntryId:
        description: |-
          PlanEntryID names this entry within the plan
          — see MigrationFileSystemModel.PlanEntryID.
        type: string
      srcConnection:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef'
    required:
    - dstConnection
    - srcConnection
    type: object
...
  github_com_cloud-barista_cm-centipede_dmdl_target-model.TargetDataMigrationModel:
    properties:
      targetDataMigrationModel:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_target-model.TargetDataMigrationProperty'
    required:
    - targetDataMigrationModel
    type: object
  github_com_cloud-barista_cm-centipede_dmdl_target-model.TargetDataMigrationProperty:
    properties:
      databases:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_target-model.MigrationDBModel'
        type: array
      fileSystems:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_target-model.MigrationFileSystemModel'
        type: array
      objectStorages:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_target-model.MigrationObjectStorageModel'
        type: array
    type: object
  github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  ? github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_dmdl_target-model_TargetDataMigrationModel
  : properties:
      data:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_target-model.TargetDataMigrationModel'
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  github_com_cloud-barista_cm-centipede_pkg_api_rest_model.PlanEntry:
    properties:
      dbmsFilter:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.DBMSFilter'
      dstConnection:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef'
      fileSystemFilter:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.FileSystemFilter'
      fileSystemStrategy:
        description: |-
          FileSystemStrategy is the transfer strategy applied to filesystem (SSH→SSH)
          tasks. Empty is normalised to "relay" (safer default: avoids agent-forward
          setup requirements). It has no effect on object-storage or DBMS tasks, where
          relay is always used / not applicable.
        enum:
        - auto
        - direct
        - relay
        type: string
      objectStorageFilter:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ObjectStorageFilter'
      srcConnection:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_common-model.ConnectionRef'
    required:
    - dstConnection
    - srcConnection
    type: object
...
  github_com_cloud-barista_cm-centipede_pkg_api_rest_model.TargetPlanReq:
    properties:
      plans:
        items:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.PlanEntry'
        minItems: 1
        type: array
      source:
        $ref: '#/definitions/github_com_cloud-barista_cm-centipede_dmdl_source-model.SourceDataMigrationModel'
    required:
    - plans
    - source
    type: object
...
host: localhost:8085
info:
  contact:
    name: Cloud-Barista Community
    url: https://github.com/cloud-barista/cm-centipede
  description: Cloud-Barista cm-centipede — Multi-cloud filesystem, object storage,
    and DBMS data migration subsystem.
  title: cm-centipede API
  version: 1.0.0
paths:
...
  /centipede/plans/target:
    post:
      consumes:
      - application/json
      description: Validates source/destination connections, checks DBMS compatibility
        and version, and returns a TargetDataMigrationModel ready for CreateMigration.
      parameters:
      - description: Target plan request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.TargetPlanReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-github_com_cloud-barista_cm-centipede_dmdl_target-model_TargetDataMigrationModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cloud-barista_cm-centipede_pkg_api_rest_model.ApiResponse-any'
      summary: Analyse target migration plan
      tags:
      - plan
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"